add scalariform and splitting all files
currently all scala code was inside a single file which made changes really hard
this moves everything into a single file
