+++ src/main/scala/play/api/libs/mailer/SMTPConfigurationModule.scala
@@ -1,0 +1,11 @@
+package play.api.libs.mailer
+
+import com.google.inject.AbstractModule
+
+class SMTPConfigurationModule extends AbstractModule {
+
+  override def configure(): Unit = {
+    bind(classOf[SMTPConfiguration]).toProvider(classOf[SMTPConfigurationProvider])
+  }
+
+}
