+++ src/main/scala/play/api/libs/mailer/SMTPConfigurationProvider.scala
@@ -1,0 +1,11 @@
+package play.api.libs.mailer
+
+import javax.inject.{ Inject, Provider }
+
+import com.typesafe.config.Config
+
+class SMTPConfigurationProvider @Inject() (config: Config) extends Provider[SMTPConfiguration] {
+  override def get(): SMTPConfiguration = {
+    SMTPConfiguration(config.getConfig("play.mailer"))
+  }
+}
